Move Wallet Contract tests to a separate file
